refs #148 Отдельный файл для свойств сортировки подсчётом
