+++ include/burst/algorithm/detail/counting_sort.hpp
@@ -1,34 +1,20 @@
 #ifndef BURST_ALGORITHM_DETAIL_COUNTING_SORT_HPP
 #define BURST_ALGORITHM_DETAIL_COUNTING_SORT_HPP
 
+#include <burst/algorithm/detail/counting_sort_traits.hpp>
 #include <burst/container/access/cback.hpp>
 #include <burst/type_traits/iterator_difference.hpp>
 #include <burst/type_traits/iterator_value.hpp>
 
 #include <algorithm>
 #include <iterator>
-#include <limits>
 #include <numeric>
-#include <type_traits>
 #include <utility>
 
 namespace burst
 {
     namespace detail
     {
-        template <typename Value, typename Map>
-        struct counting_sort_traits
-        {
-            using image_type = std::decay_t<std::result_of_t<Map(Value)>>;
-            static_assert
-            (
-                std::is_integral<image_type>::value && std::is_unsigned<image_type>::value,
-                "Сортируемые элементы должны быть отображены в целые числа."
-            );
-
-            constexpr static const auto value_range = std::numeric_limits<image_type>::max() + 1;
-        };
-
         //!     Подсчитать вхождения
         /*!
                 Для каждого сортируемого числа `n_i` подсчитывает количество его вхождений в
